Memory (un)protection in new file & multi capture triggered as tracer arg
