+++ src/memory_dump/tracer_mem.h
@@ -1,0 +1,15 @@
+#pragma once
+
+#include <sys/types.h>
+#include <stdbool.h>
+
+// LOCK
+
+void* get_protect_offset(char* executable);
+void tracer_lock_mem(pid_t pid);
+
+
+// UNLOCK
+
+bool should_unprotect(void ** start, void ** end, long long unsigned int * permissions);
+void tracer_unlock_mem(pid_t pid);
